Add classroom.rb and student.rb files
