+++ nameable.rb
@@ -3,28 +3,3 @@ class Nameable
     raise NotImplementedError, "#{self.class} has not implemented method '#{__method__}'"
   end
 end
-
-class Decorator < Nameable
-  attr_accessor :name
-
-  def initialize(nameable)
-    @nameable = nameable
-    super()
-  end
-
-  def correct_name
-    @nameable.correct_name
-  end
-end
-
-class CapitalizeDecorator < Decorator
-  def correct_name
-    @nameable.correct_name.capitalize
-  end
-end
-
-class TrimmerDecorator < Decorator
-  def correct_name
-    @nameable.correct_name.slice(0, 10)
-  end
-end
