+++ person.rb
@@ -1,7 +1,9 @@
 require './nameable'
+require './capitalize_decorator'
+require './trimmer_decorator'
 
 class Person < Nameable
-  attr_accessor :name, :age
+  attr_accessor :name, :age, :parent_permission
   attr_reader :id
 
   def initialize(age, name = 'unknown', parent_permission: true)
@@ -17,7 +19,7 @@ class Person < Nameable
   end
 
   def can_use_services?
-    if is_of_age? || @parent_permission
+    if of_age? || @parent_permission
       true
     else
       false
@@ -30,3 +32,10 @@ class Person < Nameable
     age >= 18
   end
 end
+
+# person = Person.new(22, 'maximilianus')
+# p  person.correct_name
+#   capitalizedPerson = CapitalizeDecorator.new(person)
+# p  capitalizedPerson.correct_name
+#   capitalizedTrimmedPerson = TrimmerDecorator.new(capitalizedPerson)
+# p capitalizedTrimmedPerson.correct_name
