+++ student.rb
@@ -1,12 +1,21 @@
 require './person'
 
 class Student < Person
-  def initialize(classroom, age, name = 'Unknown', parent_permission: true)
-    super(age, name, parent_permission)
+  attr_accessor :classroom
+
+  def initialize(classroom, age, name = 'Unknown', parent_permission: true, type: 'Student')
+    super(age, name, parent_permission: parent_permission)
     @classroom = classroom
+  end
+
+  def classroom=(classroom)
+    @classroom = classroom
+    classroom.students.push(self) unless classroom.students.include?(self)
   end
 
   def play_hooky
     '¯(ツ)/¯'
   end
 end
+
+
